AoR: Adding existing not saved data
In Analysis of Result it is now possible to include degenerate series with the data currently held in the configuration even if they have not been flagged to be saved. The series is composed by a single data point indicated with the conventional time stamp t=0.
The new function is activated using the shortcut Control-b from the main AoR window besides an additional option in the Add Series list.
There are also minor changes to the rest of the code:
- set as default the impossibility to delete the last instance of an object;
- changed the command to show the v[...] values in debug to avoid a bug in MacOS